# hackaton-v1/target/classes/openapi.yml
openapi: 3.0.0
info:
  title: Ecommerce API
  version: 1.0.0
  description: API para un sistema de comercio electrónico

servers:
  - url: https://api.example.com/v1

paths:
  /products:
    get:
      summary: Obtener lista de productos
      operationId: getProducts
      responses:
        '200':
          description: Lista de productos exitosa
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '500':
          description: Error del servidor

    post:
      summary: Añadir un nuevo producto
      operationId: addProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewProduct'
      responses:
        '201':
          description: Producto añadido exitosamente
        '500':
          description: Error del servidor

  /products/{productId}:
    get:
      summary: Obtener un producto por ID
      operationId: getProductById
      parameters:
        - name: productId
          in: path
          description: ID del producto a obtener
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Producto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Producto no encontrado
        '500':
          description: Error del servidor

    put:
      summary: Modificar datos de un producto por ID
      operationId: updateProductById
      parameters:
        - name: productId
          in: path
          description: ID del producto a modificar
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Producto modificado exitosamente
        '404':
          description: Producto no encontrado
        '500':
          description: Error del servidor

components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: string   # Cambiar el tipo de integer a string
        name:
          type: string
        price:
          type: number   # El tipo "number" en OpenAPI puede representar un decimal
          format: double # Indicar el formato "double" para representar números decimales
        description:
          type: string
        stock:
          type: integer

    NewProduct:
      type: object
      properties:
        name:
          type: string
        price:
          type: number   # El tipo "number" en OpenAPI puede representar un decimal
          format: double # Indicar el formato "double" para representar números decimales
        description:
          type: string
        stock:
          type: integer